refactor: extracted a small utility library to provide common rendering behavior
